Detecting correct presses.
I moved the logic part to testButtonListener.  Added a layer of time stamps to get the current and previous presses for the logic to evaluate.  It works this way, just need to practice a little with the timing for double presses.  You can fiddle with the delay times to get it working a little smoother maybe.